Struct cfile_rs::CFile

pub struct CFile {
    pub path: CString,
    // some fields omitted
}

A wrapper around C's file type. Attempts to mimic the functionality if rust's std::fs::File while still allowing complete control of all I/O operations.

Fields
path: CString
Methods

impl CFile

fn open_random_access<P: AsRef<Path>>(path: P) -> Result<CFile, Error>

Attempts to open a file in random access mode (i.e. rb+). However, unlike rb+, if the file doesn't exist, it will be created. To avoid createion, simply call CFile::open(path, "rb+"), which will return an error if the file doesn't exist.

Failure

This function will return Err for a whole bunch of reasons, the errno id will be returned as an Error::Errno(u64). For more information on what that number actually means see

fn create_file<P: AsRef<Path>>(path: &P) -> Result<(), Error>

Attempts to create a file, and then immedietly close it. If the file already exists, this function will not do anything. If the file does exist, then it will be created with no and nothing more (it will be empty).

fn open<P: AsRef<Path>>(path: P, mode: &str) -> Result<CFile, Error>

Attempt to open the file with path p.

Examples

use cfile_rs;
use cfile_rs::*;
use cfile_rs::CFile;
use cfile_rs::TRUNCATAE_RANDOM_ACCESS_MODE;
use std::str::from_utf8;

// Truncate random access mode will overwrite the old "data.txt" file if it exists.
let mut file = CFile::open("data.txt", TRUNCATAE_RANDOM_ACCESS_MODE).unwrap();

fn delete(self) -> Result<(), Error>

Deletes the file from the filesystem, and consumes the object.

Errors

On error Error::Errno(errno) is returned.

Examples

use cfile_rs;
use cfile_rs::*;
use cfile_rs::CFile;
use cfile_rs::UPDATE;
use std::str::from_utf8;

// Truncate random access mode will overwrite the old "data.txt" file if it exists.
let mut file = CFile::open("data.txt", UPDATE).unwrap();
let _ = file.write_all("Howdy folks".as_bytes());   // Write some data!
let _ = file.delete();                              // The file is gone!

fn close(self) -> Result<(), Error>

Attempts to close the file. Consumes the file as well

Errors

On error Error::Errno(errno) is returned.

unsafe fn file<'a>(&'a mut self) -> &'a mut FILE

Returns the underlying file pointer as a reference. It is returned as a reference to, in theory, prevent it from being used after the file is closed.

fn current_pos(&self) -> Result<u64, Error>

Returns the current position in the file.

Errors

On error Error::Errno(errno) is returned.